pyproject.toml
[project]
name = "ai-slide-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "langchain>=1.2.10",
    "langchain-google-genai>=4.2.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "openai>=2.24.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "python-pptx>=1.0.2",
    "uvicorn>=0.41.0",
]
